requirements.txt
Flask==2.3.2
gunicorn==21.2.0
PyJWT==2.8.0
